+++ tests/testthat/test-fits-gauss.R
@@ -1,7 +1,7 @@
 models_gauss = list(
-  # Log-link Gaussian with a CP in the mean.
+  # Log-link Gaussian with a CP in the mean and weights.
   list(
-    y ~ 1 + x,
+    y | weights(w) ~ 1 + x,
     ~ 1 + x,
     simulated = list(
       cp_1 = 93,
@@ -10,6 +10,10 @@ models_gauss = list(
       Intercept_2 = log(14),
       x_2 = -0.004,
       sigma_1 = 2
+    ),
+    newdata = data.frame(
+      x = seq(1, 200, length.out = 400),
+      w = rep(c(1, 2), 200)
     ),
     family = gaussian(link = "log"),
     chains = 2,